db fix

See merge request ppl_k03-3/backend!29

+++ routes/jadwalRouter.ts
@@ -5,6 +5,7 @@ import { Jadwal, JadwalCatalog, JadwalEdit } from "../types/jadwal";
 
 export const jadwalRouter = express.Router();
 import * as jadwalService from "../controllers/jadwalController"
+import { CustomRequest, authToken } from "../middlewares/auth";
 
 jadwalRouter.get("/viewJadwal/:jadwalId", async (request: Request, response: Response) => {
     try {
@@ -71,7 +72,7 @@ jadwalRouter.get("/listjadwal/:date", async (request: Request, response: Respons
     }
 });
 
-jadwalRouter.post("/addJadwal", [
+jadwalRouter.post("/addJadwal", authToken, [
     body("title").isString(),
     body("date").isString(),
     body("kapasitas").isNumeric(),
@@ -100,7 +101,7 @@ jadwalRouter.post("/addJadwal", [
     }
 });
 
-jadwalRouter.post("/editJadwal", [
+jadwalRouter.post("/editJadwal", authToken, [
     body("id").isNumeric(),
     body("title").isString(),
     body("date").isString(),
@@ -131,7 +132,7 @@ jadwalRouter.post("/editJadwal", [
     }
 });
 
-jadwalRouter.post("/deleteJadwal", [
+jadwalRouter.post("/deleteJadwal", authToken, [
     body("jadwalId").isNumeric(),
 ], async (request: Request, response: Response) => {
     try {
@@ -142,6 +143,15 @@ jadwalRouter.post("/deleteJadwal", [
         
         const {jadwalId} = request.body
         console.log(jadwalId)
+
+        if (!request.header('IdPlanetarium')) {
+            return response.status(403).json({ error: "Unauthorized" });
+        }
+        const idPlanetarium = parseInt(request.header('IdPlanetarium') as string);
+        const jadwalPlanetarium = await jadwalService.getjadwalById(jadwalId);
+        if (jadwalPlanetarium.planetariumId !== idPlanetarium) {
+            return response.status(403).json({ error: "Unauthorized" });
+        }
 
         await jadwalService.deleteJadwal(jadwalId);
